Improved organization of result since it's used beyond just functions
